+++ B4.Data.PostgreSQL/Services/MemoryCacheService.cs
@@ -1,11 +1,11 @@
 ﻿using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
 using B4.Models.Entities.LkEntities;
-using B4.Models.Interfaces.LkInterfaces;
+using B4.Models.RepositoryInterfaces.LkInterfaces;
 using static System.Runtime.InteropServices.JavaScript.JSType;
 using NLog;
 using System.Reflection.Metadata.Ecma335;
-using B4.Models.RepositoryInterfaces.LkInterfaces;
+
 
 namespace B4.Data.PostgreSQL.Services;